+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Add component option to lock object
 - Add footer with Trionyx and app version
 - Add changelog dialog with auto show on version change
+- Add command to generate favicon
 
 Changed
 ~~~~~~~